+++ frontend/src/bundles/auth/components/sign-in-form/sign-in-form.tsx
@@ -1,4 +1,8 @@
 import { Button } from '~/bundles/common/components/components.js';
+import {
+    ButtonSize,
+    ButtonVariant,
+} from '~/bundles/common/types/button.type.js';
 
 type Properties = {
     onSubmit: () => void;
@@ -8,7 +12,11 @@ const SignInForm: React.FC<Properties> = () => (
     <>
         <h1>Sign In</h1>
         <form>
-            <Button label="Sign in" variant="tertiary" size="medium" />
+            <Button
+                label="Sign in"
+                variant={ButtonVariant.Primary}
+                size={ButtonSize.Medium}
+            />
         </form>
     </>
 );
